refactor(kraken): fold the fees page into one card and sort the table
Filters, chart and table were spread over two cards with the filters buried in
the second one, narrowing a chart above them. They are one card now, in the
order they are read: filters, then the chart, then the table.

The chart's asset select is labelled "Chart asset" so it does not read as a
second filter next to the one above it, and its footnote is dropped — with the
two controls in view the caption was restating them.

The table's first four columns sort. An asset with no USD rate sorts last
whichever way the value column points, since an unknown value is not a small
one.
